refactor(validation): extract core contracts
Establish tools/validation as the catalog and artifact boundary for reference-consistency validation. Route the Dev/QA and reference entry points through it while keeping task_eval compatibility aliases during migration.

Fail closed when a prepared workload has no native reference runner, and document the phased path to remove the legacy task-eval implementation without changing datasets, gates, or report behavior.
