+++ tools/validation/__init__.py
@@ -1,0 +1,28 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Internal modules for TRTMC reference consistency validation."""
+
+from .artifacts import predictions_file_valid
+from .catalog import (
+    DEFAULT_MODELS_DIR,
+    DEFAULT_SUITES,
+    load_manifest_records,
+    load_structured_file,
+    load_suites,
+    resolve_suite_for_model,
+    suite_by_id,
+    suite_match_reason,
+)
+
+__all__ = [
+    "DEFAULT_MODELS_DIR",
+    "DEFAULT_SUITES",
+    "load_manifest_records",
+    "load_structured_file",
+    "load_suites",
+    "predictions_file_valid",
+    "resolve_suite_for_model",
+    "suite_by_id",
+    "suite_match_reason",
+]
